# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "SOL Project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/admin/CreationCount": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "查询邀请码邀请总数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分页",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/allCount": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "查询总注册人数",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.AllCountResponse"
                        }
                    }
                }
            }
        },
        "/admin/export-record": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "导出转账记录",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.ExportRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/export-record-team": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "导出转账记录整个团队",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.ExportTeamRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/admin/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.AdminLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/search": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "查询",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.AdminSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.AdminSearchResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "账号登录",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.LoginResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "获取用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.GetProfileResponse"
                        }
                    }
                }
            }
        },
        "/user/Invitecode": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "使用邀请码生成邀请链接",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.InviteCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.InviteCodeResponse"
                        }
                    }
                }
            }
        },
        "/user/Solsearch": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "搜索SOL地址",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.SolSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.SolSearchResponse"
                        }
                    }
                }
            }
        },
        "/user/bind/evmaddress": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "绑定evm地址",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.BindEvmAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/user/claim/reward": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "提取奖励",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.ClaimRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/user/horsh/transfer": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "Horsh交易确认",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.HorshTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/user/select": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "邀请人数查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.SelectResponse"
                        }
                    }
                }
            }
        },
        "/user/usdt/transfer": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "USDT交易确认",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.USDTTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SolProject_api_v1.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "SolProject_api_v1.AdminLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "0x2w323423"
                }
            }
        },
        "SolProject_api_v1.AdminSearchRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "code": {
                    "description": "邀请码",
                    "type": "string"
                },
                "end_time": {
                    "description": "结束时间",
                    "type": "string",
                    "example": "2024-07-01 15:03:04"
                },
                "invite_by": {
                    "description": "被谁邀请",
                    "type": "string"
                },
                "is_inviteby": {
                    "description": "是否去除被邀请",
                    "type": "boolean"
                },
                "is_not_empty_amount": {
                    "description": "SOL数量不为空",
                    "type": "boolean"
                },
                "page": {
                    "description": "\"分页\"",
                    "type": "integer"
                },
                "page_size": {
                    "description": "\"每页数量\"",
                    "type": "integer"
                },
                "start_time": {
                    "description": "开始时间",
                    "type": "string",
                    "example": "2024-06-01 15:03:04"
                }
            }
        },
        "SolProject_api_v1.AdminSearchResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/SolProject_api_v1.AdminSearchResponseDa"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "SolProject_api_v1.AllCountResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/SolProject_api_v1.AdminAllCountResponseData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "SolProject_api_v1.ExportRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "end_time": {
                    "description": "结束时间",
                    "type": "string",
                    "example": "2024-07-01 15:03:04"
                },
                "start_time": {
                    "description": "开始时间",
                    "type": "string",
                    "example": "2024-06-01 15:03:04"
                }
            }
        },
        "SolProject_api_v1.ExportTeamRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "description": "地址",
                    "type": "string"
                }
            }
        },
        "SolProject_api_v1.GetProfileResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/SolProject_api_v1.GetProfileResponseData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "SolProject_api_v1.LoginRequest": {
            "type": "object",
            "required": [
                "address",
                "login_type",
                "original_msg",
                "signature_hex"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x2w323423"
                },
                "login_type": {
                    "type": "string",
                    "enum": [
                        "sol",
                        "evm"
                    ]
                },
                "original_msg": {
                    "type": "string"
                },
                "signature_hex": {
                    "type": "string"
                }
            }
        },
        "SolProject_api_v1.LoginResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/SolProject_api_v1.LoginResponseData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "SolProject_api_v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "SolProject_api_v1.SolSearchRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Evm地址",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}